Use the refresh token to obtain a fresh access token each time
The Google OAuth tokens set an expiry, however seem to work long after
that expiry has passed. They do, however, eventually stop working, so
we need to use the refresh token to get a new valid token.

As we aren't able to store data between pipeline runs, we can't check the
expiry of the token we have, and need to get a new token each time.