Add Secondary Staging Smoke Tests
Add Secondary Staging smoke-tests.
Add new `SUBDOMAIN` environment variable to smoke-test to make it easier to adapt the tests to different environments in future (e.g. production, staging, development etc)
Comment out Slack notifications in Secondary Staging environment for now. We will re-enable them when work on the Secondary Staging environment is complete.

We need these to ensure that the apps and infracstructure in the Secondary Staging environment are functioning correctly.