Use the deploy- secrets for image access for the staging tests
This means that the ECR repository in the production account is used,
rather than the one in the staging account.

Because only the production repository was being updated by the
pipeline, the latest image in the staging repository was quite out
date, and eventually this caused the smoke tests to fail (because out
of date code was being used).