What Remove references to "secondary staging"
Why
The new separate staging environment is now live, and the old staging
environment has been deleted. Removing references to "secondary" to avoid later
confusion. The secret this refers to will be updated just before this is merged
into the main branch.

Link to Trello card (if applicable): https://trello.com/c/r6dKgFUR/1542-story-integrate-existing-main-concourse-pipeline-with-secondary-staging-so-it-will-push-to-secondary-staging