Fix intermittent failuers in smoke tests.
Every now and then the smoke tests fail. An Eapol test would be performed
for a username and a password and this would fail against all radius servers.
The username would appear in the sessions table in the sessions DB as
failed. The username however would be absent in the userdetails table in
the users DB. Further investigations uncovered it was present in th
user details table in the staging environment.

This leads us to believe emails are erroneously read from staging
instead of production, leading to this error.

This fix filters messages from a particular environment, thereby fixing
the problem