Implement the SMS signup journey tests
This test uses a Notify service account called GovWifi-Smoketest.
This account is used to send and receive SMS messages similar
to how a regular GovWifi user would

Synchronisation is handled by reading the first SMS message from
the GovWIfi-Smoketest account, then sending a 'go' message to start
the signup process. The GovWifi-Smoketest account is then polled for
the first message that is different from the one read earlier.